Move code to index.js, make index.js main
This will enable this module to be consumed on `unpkg.org` and other
systems that happily support its UMD export style, but do not resolve
the require() statement previously in the file.
