+++ package.json
@@ -33,7 +33,6 @@
     "style": "colorbrewer.css",
     "files": [
         "colorbrewer.css",
-        "colorbrewer.js",
         "index.js"
     ],
     "main": "index.js"
